test: merge StreamHeaderBlockSizeTests into StreamDecoderTests
Both hand-built minimal Vorbis header packets via identical
BitWriter/ByteArrayPacket/FakeHeaderPacketProvider/ConstructAndCapture
helpers after the fix/mdct-small-block-and-perf rebase. Fold the
block-size theory tests in and dedupe ValidStreamHeader onto the new
StreamHeaderWithBlockSizes(block0Exp, block1Exp) helper instead of
keeping two copies of the same header-byte layout.
